requirements.txt
flask
numpy
pillow
numba
flask_login
sqlalchemy
sqlalchemy_serializer